改了一些地方的变量声明

+++ TripleBit/MMapBuffer.h
@@ -28,7 +28,7 @@ public:
 	unsigned long long get_length() { return size;}
 	char * get_address() const { return (char*)mmap_addr; }
 
-	virtual Status resize(size_t new_size,bool clear);
+	virtual Status resize(unsigned long	long new_size,bool clear);
 	virtual void   memset(char value);
 
 	MMapBuffer(const char* filename, size_t initSize);
